// src/app/Enums/elemento-procesador.ts
export enum ElementoProcesador {
  PC = 'PC',
  MAR = 'MAR',
  MBR = 'MBR',
  IR = 'IR',
  ALU = 'ALU',
  MEMORIA = 'Memoria',
  ALMACEN_GENERAL = 'Almacen general',
  BUS_DATOS = 'Bus datos',
  BUS_DIRECCIONES = 'Bus direcciones',
  BUS_CONTROL = 'Bus control',
  UNIDAD_CONTROL = 'Unidad de control',
}